apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
email-validator==2.2.0
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1
